+++ src/cgame/cg_marks.c
@@ -51,7 +51,7 @@ void CG_InitMarkPolys( void )
 {
   int   i;
 
-  memset( cg_markPolys, 0, sizeof( cg_markPolys ) );
+  memset( cg_markPolys, 0, (int)sizeof( cg_markPolys ) );
 
   cg_activeMarkPolys.nextMark = &cg_activeMarkPolys;
   cg_activeMarkPolys.prevMark = &cg_activeMarkPolys;
@@ -106,7 +106,7 @@ markPoly_t *CG_AllocMark( void )
   le = cg_freeMarkPolys;
   cg_freeMarkPolys = cg_freeMarkPolys->nextMark;
 
-  memset( le, 0, sizeof( *le ) );
+  memset( le, 0, (int)sizeof( *le ) );
 
   // link into the active list
   le->nextMark = cg_activeMarkPolys.nextMark;
@@ -224,7 +224,7 @@ void CG_ImpactMark( qhandle_t markShader, const vec3_t origin, const vec3_t dir,
     mark->color[ 1 ] = green;
     mark->color[ 2 ] = blue;
     mark->color[ 3 ] = alpha;
-    memcpy( mark->verts, verts, mf->numPoints * sizeof( verts[ 0 ] ) );
+    memcpy( mark->verts, verts, mf->numPoints * (int)sizeof( verts[ 0 ] ) );
     markTotal++;
   }
 }
